refactor some calibration plguin code
