+++ pupil_src/shared_modules/calibration_routines/natural_features_calibration.py
@@ -19,7 +19,7 @@ import audio
 
 
 from pyglui import ui
-from plugin import Calibration_Plugin
+from calibration_plugin_base import Calibration_Plugin
 
 #logging
 import logging
@@ -35,7 +35,6 @@ class Natural_Features_Calibration(Calibration_Plugin):
         self.point = None
         self.count = 0
         self.detected = False
-        self.active = False
         self.pos = None
         self.r = 40.0 # radius of circle displayed
         self.ref_list = []
@@ -67,9 +66,10 @@ class Natural_Features_Calibration(Calibration_Plugin):
 
     def toggle(self,_=None):
         if self.active:
-            self.stop()
+            self.notify_all({'subject':'calibration.should_stop'})
         else:
-            self.start()
+            self.notify_all({'subject':'calibration.should_start'})
+
 
     def start(self):
         audio.say("Starting Calibration")
@@ -116,7 +116,7 @@ class Natural_Features_Calibration(Calibration_Plugin):
 
             #always save pupil positions
             for p_pt in recent_pupil_positions:
-                if p_pt['confidence'] > self.g_pool.pupil_confidence_threshold:
+                if p_pt['confidence'] > self.pupil_confidence_threshold:
                     self.pupil_list.append(p_pt)
 
             if self.count:
